+++ bin/config-yargs.js
@@ -1,4 +1,4 @@
-const optionsSchema = require("webpack/schemas/WebpackOptions.json");
+const optionsSchema = require("./optionsSchema.json");
 const CONFIG_GROUP = "Config options:";
 const BASIC_GROUP = "Basic options:";
 const MODULE_GROUP = "Module options:";
@@ -6,7 +6,6 @@ const OUTPUT_GROUP = "Output options:";
 const ADVANCED_GROUP = "Advanced options:";
 const RESOLVE_GROUP = "Resolving options:";
 const OPTIMIZE_GROUP = "Optimizing options:";
-const INIT_GROUP = "Initialization:";
 
 module.exports = function(yargs) {
 	yargs
@@ -15,52 +14,6 @@ module.exports = function(yargs) {
 		.version()
 		.alias("version", "v")
 		.options({
-			init: {
-				type: "boolean",
-				describe:
-					"Initializes a new webpack configuration or loads a" +
-					"\n" +
-					"addon if specified",
-				group: INIT_GROUP
-			},
-			migrate: {
-				type: "boolean",
-				describe:
-					"Migrate your webpack configuration from webpack 1 to webpack 2",
-				group: INIT_GROUP
-			},
-			add: {
-				type: "boolean",
-				describe: "Adds a webpack component to your configuration file",
-				group: INIT_GROUP
-			},
-			/*
-			remove: {
-				type: "boolean",
-				describe: "Removes a webpack component to your configuration file",
-				group: INIT_GROUP
-			},
-			update: {
-				type: "boolean",
-				describe: "Updates a webpack component to your configuration file",
-				group: INIT_GROUP
-			},
-			make: {
-				type: "boolean",
-				describe: "Converts a webpack configuration to a makefile",
-				group: INIT_GROUP
-			},
-			*/
-			"generate-loader": {
-				type: "boolean",
-				describe: "Generates a new webpack loader project",
-				group: INIT_GROUP
-			},
-			"generate-plugin": {
-				type: "boolean",
-				describe: "Generates a new webpack plugin project",
-				group: INIT_GROUP
-			},
 			config: {
 				type: "string",
 				describe: "Path to the config file",
@@ -140,14 +93,16 @@ module.exports = function(yargs) {
 			},
 			"output-filename": {
 				type: "string",
-				describe: optionsSchema.definitions.output.properties.filename.description,
+				describe:
+					optionsSchema.definitions.output.properties.filename.description,
 				group: OUTPUT_GROUP,
 				defaultDescription: "[name].js",
 				requiresArg: true
 			},
 			"output-chunk-filename": {
 				type: "string",
-				describe: optionsSchema.definitions.output.properties.chunkFilename.description,
+				describe:
+					optionsSchema.definitions.output.properties.chunkFilename.description,
 				group: OUTPUT_GROUP,
 				defaultDescription:
 					"filename with [id] instead of [name] or [id] prefixed",
@@ -155,25 +110,30 @@ module.exports = function(yargs) {
 			},
 			"output-source-map-filename": {
 				type: "string",
-				describe: optionsSchema.definitions.output.properties.sourceMapFilename.description,
+				describe:
+					optionsSchema.definitions.output.properties.sourceMapFilename
+						.description,
 				group: OUTPUT_GROUP,
 				requiresArg: true
 			},
 			"output-public-path": {
 				type: "string",
-				describe: optionsSchema.definitions.output.properties.publicPath.description,
+				describe:
+					optionsSchema.definitions.output.properties.publicPath.description,
 				group: OUTPUT_GROUP,
 				requiresArg: true
 			},
 			"output-jsonp-function": {
 				type: "string",
-				describe: optionsSchema.definitions.output.properties.jsonpFunction.description,
+				describe:
+					optionsSchema.definitions.output.properties.jsonpFunction.description,
 				group: OUTPUT_GROUP,
 				requiresArg: true
 			},
 			"output-pathinfo": {
 				type: "boolean",
-				describe: optionsSchema.definitions.output.properties.pathinfo.description,
+				describe:
+					optionsSchema.definitions.output.properties.pathinfo.description,
 				group: OUTPUT_GROUP
 			},
 			"output-library": {
@@ -184,7 +144,8 @@ module.exports = function(yargs) {
 			},
 			"output-library-target": {
 				type: "string",
-				describe: optionsSchema.definitions.output.properties.libraryTarget.description,
+				describe:
+					optionsSchema.definitions.output.properties.libraryTarget.description,
 				choices: optionsSchema.definitions.output.properties.libraryTarget.enum,
 				group: OUTPUT_GROUP,
 				requiresArg: true
@@ -235,18 +196,24 @@ module.exports = function(yargs) {
 			"watch-stdin": {
 				type: "boolean",
 				alias: "stdin",
-				describe: optionsSchema.properties.watchOptions.properties.stdin.description,
+				describe:
+					optionsSchema.properties.watchOptions.properties.stdin.description,
 				group: ADVANCED_GROUP
 			},
 			"watch-aggregate-timeout": {
-				describe: optionsSchema.properties.watchOptions.properties.aggregateTimeout.description,
-				type: optionsSchema.properties.watchOptions.properties.aggregateTimeout.type,
+				describe:
+					optionsSchema.properties.watchOptions.properties.aggregateTimeout
+						.description,
+				type:
+					optionsSchema.properties.watchOptions.properties.aggregateTimeout
+						.type,
 				group: ADVANCED_GROUP,
 				requiresArg: true
 			},
 			"watch-poll": {
 				type: "string",
-				describe: optionsSchema.properties.watchOptions.properties.poll.description,
+				describe:
+					optionsSchema.properties.watchOptions.properties.poll.description,
 				group: ADVANCED_GROUP
 			},
 			hot: {
@@ -267,13 +234,15 @@ module.exports = function(yargs) {
 			},
 			"resolve-alias": {
 				type: "string",
-				describe: optionsSchema.definitions.resolve.properties.alias.description,
+				describe:
+					optionsSchema.definitions.resolve.properties.alias.description,
 				group: RESOLVE_GROUP,
 				requiresArg: true
 			},
 			"resolve-extensions": {
 				type: "array",
-				describe: optionsSchema.definitions.resolve.properties.alias.description,
+				describe:
+					optionsSchema.definitions.resolve.properties.alias.description,
 				group: RESOLVE_GROUP,
 				requiresArg: true
 			},
@@ -289,13 +258,16 @@ module.exports = function(yargs) {
 				requiresArg: true
 			},
 			"optimize-min-chunk-size": {
-				describe: optionsSchema.properties.optimization.properties.splitChunks.oneOf[1].properties.minSize.description,
+				describe:
+					optionsSchema.properties.optimization.properties.splitChunks.oneOf[1]
+						.properties.minSize.description,
 				group: OPTIMIZE_GROUP,
 				requiresArg: true
 			},
 			"optimize-minimize": {
 				type: "boolean",
-				describe: optionsSchema.properties.optimization.properties.minimize.description,
+				describe:
+					optionsSchema.properties.optimization.properties.minimize.description,
 				group: OPTIMIZE_GROUP
 			},
 			prefetch: {
@@ -346,6 +318,5 @@ module.exports = function(yargs) {
 					"shortcut for --optimize-minimize --define process.env.NODE_ENV=\"production\"",
 				group: BASIC_GROUP
 			}
-		})
-		.strict();
+		});
 };