+++ packages/migrate/uglifyJsPlugin/__testfixtures__/uglifyJsPlugin-0.input.js
@@ -1,0 +1,7 @@
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
+
+module.exports = {
+    plugins: [
+        new UglifyJsPlugin()
+    ]
+}