+++ test/binCases/config-type/array/stdin.js
@@ -7,7 +7,7 @@ module.exports = function testAssertions(code, stdout, stderr) {
 	expect(stdout[1]).toContain("Version: ");
 	expect(stdout[2]).toContain("Child");
 	expect(stdout[7]).toContain("entry-a.bundle.js");
-	expect(stdout[10]).toContain("Child");
-	expect(stdout[15]).toContain("entry-b.bundle.js");
+	expect(stdout[9]).toContain("Child");
+	expect(stdout[14]).toContain("entry-b.bundle.js");
 	expect(stderr).toHaveLength(0);
 };