+++ test/binCases/errors/issue-5576/stdin.js
@@ -5,7 +5,7 @@ module.exports = function testAssertions(code, stdout, stderr) {
 	expect(stdout[0]).toContain("Hash: ");
 	expect(stdout[1]).toContain("Version: ");
 	expect(stdout[2]).toContain("Time: ");
-	expect(stdout[6]).toContain("bundle.js");
+	expect(stdout[5]).toContain("bundle.js");
 
 	expect(stderr).toHaveLength(0);
 };