+++ test/binCases/output/output-argument/stdin.js
@@ -4,6 +4,6 @@ module.exports = function testAssertions(code, stdout, stderr) {
 	expect(code).toBe(0);
 	expect(stdout).toEqual(expect.anything());
 	expect(stdout[5]).toContain("bundle.js");
-	expect(stdout[7]).toMatch(/index\.js.*\{0\}/);
+	expect(stdout[6]).toMatch(/index\.js.*\{0\}/);
 	expect(stderr).toHaveLength(0);
 };